requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==21.2.0
awscli==1.27.144
backoff==2.2.1
blinker==1.4
botocore==1.29.144
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
cmake==3.26.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.5.7
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duckdb==0.8.0
emoji==1.2.0
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==1.1.2
h11==0.14.0
hnswlib==0.7.0
httplib2==0.20.2
httptools==0.5.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==3.2.0
keyring==23.5.0
kss==4.5.3
langchain==0.0.187
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.5
loguru==0.7.0
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==1.4.6
neovim==0.3.1
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pecab==1.0.8
Pillow==9.5.0
pinecone-client==2.2.1
pluggy==1.0.0
posthog==3.0.1
protobuf==4.23.2
pyarrow==12.0.0
pyasn1==0.5.0
pybind11==2.9.2
pydantic==1.10.8
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
pypdf==3.9.0
pyrsistent==0.18.1
pytest==7.3.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-mecab-kor==1.2.8
pytz==2023.3
PyYAML==5.4.1
regex==2023.5.5
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
systemd-python==234
tenacity==8.2.2
texttable==1.6.4
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.19.0
websocket-client==1.2.3
websockets==11.0.3
yarl==1.9.2
zipp==1.0.0
zstandard==0.21.0
